convert notes to float

+++ examples/examples-stk/streams-stk_sine-audiokit/streams-stk_sine-audiokit.ino
@@ -31,7 +31,7 @@ void play() {
   if (millis()>timeout){
     if (active){
       // play note for 800 ms
-      freq = notes_array[random(sizeof(notes_array)/sizeof(uint16_t))];
+      freq = notes_array[random(sizeof(notes_array)/sizeof(float))];
 
       Serial.print("playing ");
       Serial.println(freq);
